// src/components/pages/TourDetailsScreen/type.ts
export interface TourDetailsType {
  agencyName: string;
  departureDate: string;
  departureTime: string;
  destination: string;
  hasBirthCertificate: boolean;
  hasDestinationVisa: boolean;
  hasDrivingCertificate: boolean;
  hasHealthEnsuranceCard: boolean;
  hasNationalCard: boolean;
  hasOff: boolean;
  hasPassport: boolean;
  hasPrivacyDocument: boolean;
  images: string[];
  offAmount: string;
  origin: string;
  price: number;
  returnDate: string;
  returnTime: string;
  services: string;
  title: string;
  tourCode: string;
}